Add NUC 6CAYH UEFI hardware configuration
Required to make SBS run fast on this target, SHA512 signatures with
BIOS legacy boot are terribly slow. The reason is not yet clear. The
same image runs fast e.g. on the Lenovo T430s/T440s/T480s notebooks.

GRUB's PGP with multiboot shows the same effect, so this seems to be a
GRUB2 bug.