Mugenspec: Add annotation to IOMMUs instance
Accept warning regarding IOMMUs type not being suitable for aliasing
via address clause, which is emitted due to the record containing holes.
This use case is safe, because we are interacting with a memory mapped
device and only accessing the specified registers through the record
with an appropriate representation clause.

Note: Disabling warnings is necessary, since the message is only emitted
      by recent GNATprove versions. SPARK tools that do not raise the
      annotated warning will in turn abort with an error as in that case
      there is no matching warning.