Libdebuglog: Accept warning for log channel instance
Accept warning that the shared memory channel type is not suitable
for aliasing via address clause. Currently, there is a limitation in
GNATprove, which makes is necessary to declare the Size aspect of all
types used in composite types. However, this does not work well in
combination with generics, where the size of the effective memory
channel is different for each generic instance, see also ticket
[T611-019].

Note: Disabling warnings is necessary, since the message is only emitted
      by recent GNATprove versions. SPARK tools that do not raise the
      annotated warning will in turn abort with an error as in that case
      there is no matching warning.